Move malloc free many size to tests
